[REF] mail: move call thread code to channel
