+++ addons/mail/static/src/discuss/call/public_web/discuss_app/client_action_patch.js
@@ -28,7 +28,7 @@ patch(DiscussClientAction.prototype, {
         if (
             hasFullScreenUrl &&
             this.store.discuss.thread?.channel.default_display_mode === "video_full_screen" &&
-            this.store.discuss.thread.rtc_session_ids.length > 0
+            this.store.discuss.thread.channel?.hasRtcSessionActive
         ) {
             this.joinCallWithDefaultSettings();
         }
