+++ addons/mail/static/src/discuss/call/public_web/discuss_app/sidebar/call_indicator.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
     <t t-name="mail.DiscussSidebarCallIndicator">
-        <div t-if="props.channel.rtc_session_ids.length > 0" title="Ongoing call" class="o-mail-DiscussSidebarCallIndicator fa-fw rounded-circle fa fa-volume-up" style="padding-top: 1px; padding-bottom: 1px;" t-att-class="{
+        <div t-if="props.channel.hasRtcSessionActive" title="Ongoing call" class="o-mail-DiscussSidebarCallIndicator fa-fw rounded-circle fa fa-volume-up" style="padding-top: 1px; padding-bottom: 1px;" t-att-class="{
             'o-discuss-inCallIconColor': props.channel.eq(rtc.channel),
             'opacity-50': !props.channel.eq(rtc.channel),
         }"/>
