+++ addons/mail/static/src/discuss/core/public_web/discuss_content_patch.xml
@@ -3,7 +3,7 @@
     <t t-inherit="mail.DiscussContent" t-inherit-mode="extension">
         <xpath expr="//Thread" position="before">
             <PipBanner t-if="rtc.isPipMode and thread.channel?.eq(rtc.channel)"/>
-            <Call t-elif="thread.showCallView" channel="thread.channel"/>
+            <Call t-elif="thread.channel?.showCallView" channel="thread.channel"/>
         </xpath>
     </t>
 </templates>
